requirement.txt
absl-py==1.3.0
addict==2.4.0
ansi2html==1.8.0
antlr4-python3-runtime==4.9.3
attrs==22.2.0
backports.cached-property==1.0.2
bleach==5.0.1
bokeh==2.4.3
CacheControl==0.12.11
cdflib==0.3.20
cffi==1.15.1
charset-normalizer==2.1.1
cleo==2.0.1
click==8.1.5
ConfigArgParse==1.5.5
crashtest==0.4.1
cryptography==39.0.0
cycler==0.11.0
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2021.10.0
debugpy==1.6.7
distlib==0.3.6
dulwich==0.20.50
easydict==1.10
einops==0.6.0
entrypoints==0.4
exceptiongroup==1.1.0
fastjsonschema==2.17.1
filelock==3.8.2
filterpy==1.4.5
Flask==2.2.5
flatbuffers==2.0.7
fonttools==4.38.0
freetype-py==2.3.0
future==0.18.2
html5lib==1.1
huggingface-hub==0.11.1
hydra-core==1.3.1
idna==3.4
imageio==2.23.0
importlib-metadata==4.13.0
importlib-resources==5.10.2
iniconfig==2.0.0
ipykernel==6.16.2
ipywidgets==8.0.7
itsdangerous==2.1.2
jaraco.classes==3.2.3
jax==0.3.25
jaxlib==0.3.25+cuda11.cudnn82
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-widgets==3.0.8
keyring==23.13.1
llvmlite==0.39.1
lockfile==0.12.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-service==2.4.0
more-itertools==9.0.0
msgpack==1.0.4
munkres==1.1.4
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.6.3
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
omegaconf==2.3.0
open3d==0.17.0
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
pandas==1.1.5
panel==0.14.2
param==1.12.3
Pillow==9.4.0
pkginfo==1.9.5
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
plotly==5.15.0
pluggy==1.0.0
ply==3.11
poetry==1.3.1
poetry-core==1.4.0
poetry-plugin-export==1.2.0
protobuf==3.20.3
psutil==5.9.5
pycparser==2.21
pyct==0.4.8
pyglet==1.5.9
pymp==0.0.6
pymp-pypi==0.4.5
PyOpenGL==3.1.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyquaternion==0.9.9
pyrender==0.1.45
pyrsistent==0.19.3
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.7
pyviz-comms==2.2.1
PyYAML==6.0
pyzmq==25.1.0
rapidfuzz==2.13.7
requests==2.28.1
requests-toolbelt==0.10.1
retrying==1.3.4
scikit-learn==1.0.2
SecretStorage==3.3.3
shellingham==1.5.0.post1
six==1.16.0
smplx==0.1.28
tenacity==8.2.2
tensorboardX==2.6
threadpoolctl==3.1.0
timm==0.4.5
tomli==2.0.1
tomlkit==0.11.6
torch==1.12.1+cu116
torch-dct==0.1.6
torchaudio==0.12.1+cu116
torchgeometry==0.1.2
torchvision==0.13.1+cu116
tqdm==4.64.1
transform3d==0.0.4
transforms3d==0.4.1
trimesh==3.17.1
trove-classifiers==2022.12.22
typing_extensions==4.4.0
urllib3==1.26.13
virtualenv==20.17.1
webencodings==0.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.8
